Add files via upload
In this machine learning model we want to reliably predict if a vehicle we buy is going to be sold in a period of time less than 2 months. The COLUMN MONTHS in the excel VENDIDOS1MESHISTORICOml2.xlsx HAS 2 VALUES:
1 IF WE SELL THE VEHICLE IN LESS THAN 2 MONTHS
2 WHEN WE EXCEED THAT PERIOD.
We will use a DECISION TREE MODEL and three K nearest neighbor with k=3 k=7 and k=15.